orangepi/requirements.txt
# requirements for the orangepi flight-tracker application
requests
python-dotenv